+++ src/models/character/keybinding/model.rs
@@ -1,16 +1,15 @@
-use std::time::SystemTime;
-
 use crate::db::schema::keybindings;
 use diesel::prelude::*;
+use std::time::SystemTime;
 
 #[derive(Clone, Identifiable, Queryable, Insertable, AsChangeset)]
 #[diesel(table_name = keybindings)]
 pub struct Keybinding {
     pub id: i32,
     pub char_id: i32,
-    pub key: i16,
+    pub key: i32,
     pub bind_type: i16,
-    pub action: i16,
+    pub action: i32,
     pub created_at: Option<SystemTime>,
     pub updated_at: Option<SystemTime>,
 }
@@ -19,9 +18,9 @@ pub struct Keybinding {
 #[diesel(table_name = keybindings)]
 pub struct NewKeybinding {
     pub char_id: i32,
-    pub key: i16,
+    pub key: i32,
     pub bind_type: i16,
-    pub action: i16,
+    pub action: i32,
 }
 
 #[derive(Debug, Clone, Copy)]
@@ -38,7 +37,7 @@ pub enum KeybindType {
 }
 
 impl Keybinding {
-    pub fn empty(char_id: i32, key: i16) -> Self {
+    pub fn empty(char_id: i32, key: i32) -> Self {
         Self {
             id: 0,
             char_id,
