+++ src/models/world/error.rs
@@ -4,10 +4,10 @@ use thiserror::Error;
 #[derive(Debug, Error)]
 pub enum WorldError {
     #[error("Requested world was not found in world model layer: {0}")]
-    NotFound(i16),
+    NotFound(i8),
 
     #[error("No world selected for in world model layer: {0}")]
-    NotSelected(i32),
+    NotSelected(i8),
 
     #[error("Config error in world model layer")]
     ConfigError(#[from] ConfigError),
