+++ src/models/world/model.rs
@@ -4,7 +4,7 @@ use diesel::prelude::*;
 use std::time::SystemTime;
 
 pub struct World {
-    pub id: i16,
+    pub id: i8,
     pub name: String,
     pub flag: i8,
     pub event_message: String,
@@ -15,7 +15,7 @@ pub struct World {
 #[diesel(table_name = character_limits)]
 pub struct CharacterLimit {
     pub acc_id: i32,
-    pub char_max: i32,
+    pub char_max: i16,
     pub world_id: i16,
     pub created_at: Option<SystemTime>,
     pub updated_at: Option<SystemTime>,
